Moves Exceptions into separate namespace
- Moves Exceptions from Annotation.php into separate files
- Updates usage in Annotation.php and AnnotationTest.php to reflect this change
