+++ src/QCheck/Exceptions/AmbiguousTypeAnnotationException.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace QCheck\Exceptions;
+
+class AmbiguousTypeAnnotationException extends AnnotationException {}
