+++ src/QCheck/Exceptions/AnnotationException.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace QCheck\Exceptions;
+
+class AnnotationException extends \Exception {}
