+++ src/QCheck/Exceptions/DuplicateGeneratorException.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace QCheck\Exceptions;
+
+class DuplicateGeneratorException extends AnnotationException {}
