+++ src/QCheck/Exceptions/MissingTypeAnnotationException.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace QCheck\Exceptions;
+
+class MissingTypeAnnotationException extends AnnotationException {}
