+++ src/QCheck/Exceptions/NoGeneratorAnnotationException.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace QCheck\Exceptions;
+
+class NoGeneratorAnnotationException extends AnnotationException {}
